re-shuffle locations
