+++ task-sdk/src/airflow/sdk/definitions/deadline.py
@@ -1,0 +1,66 @@
+# Licensed to the Apache Software Foundation (ASF) under one
+# or more contributor license agreements.  See the NOTICE file
+# distributed with this work for additional information
+# regarding copyright ownership.  The ASF licenses this file
+# to you under the Apache License, Version 2.0 (the
+# "License"); you may not use this file except in compliance
+# with the License.  You may obtain a copy of the License at
+#
+#   http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing,
+# software distributed under the License is distributed on an
+# "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+# KIND, either express or implied.  See the License for the
+# specific language governing permissions and limitations
+# under the License.
+from __future__ import annotations
+
+from datetime import datetime
+
+
+class DeadlineReference:
+    """
+    The public interface class for all DeadlineReference options.
+
+    This class provides a unified interface for working with Deadlines, supporting both
+    calculated deadlines (which fetch values from the database) and fixed deadlines
+    (which return a predefined datetime).
+
+    ------
+    Usage:
+    ------
+
+    1. Example deadline references:
+       fixed = DeadlineReference.FIXED_DATETIME(datetime(2025, 5, 4))
+       logical = DeadlineReference.DAGRUN_LOGICAL_DATE
+       queued = DeadlineReference.DAGRUN_QUEUED_AT
+
+    2. Using in a DAG:
+       DAG(
+           dag_id='dag_with_deadline',
+           deadline=DeadlineAlert(
+               reference=DeadlineReference.DAGRUN_LOGICAL_DATE,
+               interval=timedelta(hours=1),
+               callback=hello_callback,
+           )
+       )
+
+    3. Evaluating deadlines will ignore unexpected parameters:
+       # For deadlines requiring parameters:
+           deadline = DeadlineReference.DAGRUN_LOGICAL_DATE
+           deadline.evaluate_with(dag_id=dag.dag_id)
+
+       # For deadlines with no required parameters:
+           deadline = DeadlineReference.FIXED_DATETIME(datetime(2025, 5, 4))
+           deadline.evaluate_with()
+    """
+
+    from airflow.models.deadline import ReferenceModels
+
+    DAGRUN_LOGICAL_DATE: ReferenceModels.BaseDeadlineReference = ReferenceModels.DagRunLogicalDateDeadline()
+    DAGRUN_QUEUED_AT: ReferenceModels.BaseDeadlineReference = ReferenceModels.DagRunQueuedAtDeadline()
+
+    @classmethod
+    def FIXED_DATETIME(cls, datetime: datetime) -> ReferenceModels.BaseDeadlineReference:
+        return cls.ReferenceModels.FixedDatetimeDeadline(datetime)
